Added cache functions to also have operations on the btree if possible

+++ src/Configuration.h
@@ -17,9 +17,9 @@ namespace Configuration
 
     struct Configuration
     {
-        int buffer_size = 5;                  /// size of database availabile in memory
-        int record_count = 1000;              /// number of elements saved during workload
-        int operation_count = 1000;           /// operation count during workload
+        uint64_t buffer_size = 5;                  /// size of database availabile in memory
+        uint64_t record_count = 1000;              /// number of elements saved during workload
+        uint64_t operation_count = 1000;           /// operation count during workload
         std::string distribution = "uniform"; /// distribution during workload
         double insert_proportion = 0;         /// proportion of inserts during workload
         double read_proportion = 0.5;         /// proportion of read during workload
@@ -27,7 +27,7 @@ namespace Configuration
         double scan_proportion = 0;           /// proportion of scan during workload
         double delete_proportion = 0;         /// proportion of delete during workload
         bool cache = false;                   /// if caching is enabled
-        int radix_tree_size = 104857600;      /// Size of the cache, default here is 100 MB
+        uint64_t radix_tree_size = 104857600;      /// Size of the cache, default here is 100 MB
         bool measure_per_operation = false;   /// Either measure throughput or individual operations which gives the percentiles etc.
         bool benchmark = false;               /// whether benchmarking is enabled or not, only applicable for run config
         bool run_workload = false;            /// if a workload or a run config should be run
